src/third_party/blink/web_tests/http/tests/origin_trials/webexposed/webcomponentsv0-basic-import-enabled-insecure-context.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- NOTE: The token must match the UNAUTHENTICATED_ORIGIN used in the test.
  Generate this token with the command:
generate_token.py http://example.test:8000 WebComponentsV0 --expire-timestamp=2000000000
-->
<meta http-equiv="origin-trial" content="ApvQmWyEJ1hMIUp80AK+wjmYaGlmF2UA9MSWCt5vej5J+gj5XsYaAyyczK5df2RvdvrgZ8YloMdZRXp3v1pi9AoAAABaeyJvcmlnaW4iOiAiaHR0cDovL2V4YW1wbGUudGVzdDo4MDAwIiwgImZlYXR1cmUiOiAiV2ViQ29tcG9uZW50c1YwIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9"/>

<link id="target" rel="import" href="resources/html-imports-hello.html">  

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/get-host-info.js"></script>
<script>
  if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
    window.location = get_host_info().UNAUTHENTICATED_ORIGIN +
      window.location.pathname;
  } else {
    test(() => {
        assert_equals(target.import.querySelector('h1').textContent, 'Hello');
      }, 'Basic html file should be imported in insecure context');
  }
</script>
